+++ packages/apollo-link-schema/CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change log
 
 ### vNEXT
+- pass context from operation into execute
+- remove `context` constructor param
 
 ### 1.0.2
 - handle synchronous result of `execute` **Fixes #351**
-- pass context from operation into execute
 
 ### 1.0.1
 - export options as named interface [TypeScript]